+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+Features:
+- Add `link` props to `Field.Label`
+- Add `Field.Password` component
 
 ## [1.21.0] - 2019-03-11