+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `customSize` props to `HorizontalStroke`.
 - Feature: Add `HorizontalStroke` size CSS: `TINY`, `DEFAULT`, `BIG` and `HUGE`.
 - Feature: Add `IconList`.
+- Feature: Add empty `Prev` and `Next` container to `NavTabCarousel` when
+  `NavTabCarousel.Prev` or `NavTabCarousel.Next` are empty.
 
 ## [1.21.0] - 2019-03-11